requirements.txt
Django==1.8
django-rest-auth==0.6.0
djangorestframework==3.3.2
django-allauth==0.24.1
django-cors-headers==1.0.0
django-oauth-toolkit==0.10.0
Fabric==1.10.2